+++ PanoView/Panorama.storyboard
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16097" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="BRP-0H-eOd">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Panorama View Controller-->
+        <scene sceneID="xgp-Sg-spz">
+            <objects>
+                <viewController id="BRP-0H-eOd" customClass="PanoramaViewController" customModule="PanoView" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="Dhn-mi-ndQ">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <arscnView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="eg0-fs-3qd">
+                                <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                            </arscnView>
+                            <slider opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" value="0.5" minValue="0.0" maxValue="15" translatesAutoresizingMaskIntoConstraints="NO" id="dBL-fd-hIf">
+                                <rect key="frame" x="202" y="84" width="199" height="31"/>
+                                <constraints>
+                                    <constraint firstAttribute="width" constant="195" id="3O7-6x-AF0"/>
+                                    <constraint firstAttribute="height" constant="30" id="QCo-IP-NAS"/>
+                                </constraints>
+                                <connections>
+                                    <action selector="zoomSlider:" destination="BRP-0H-eOd" eventType="valueChanged" id="zOy-1X-Oyw"/>
+                                </connections>
+                            </slider>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstItem="eg0-fs-3qd" firstAttribute="leading" secondItem="Dhn-mi-ndQ" secondAttribute="leading" id="CBa-pB-SQi"/>
+                            <constraint firstItem="dBL-fd-hIf" firstAttribute="top" secondItem="FOa-2C-Uzo" secondAttribute="top" constant="40" id="Rhb-Zo-ga6"/>
+                            <constraint firstItem="eg0-fs-3qd" firstAttribute="trailing" secondItem="Dhn-mi-ndQ" secondAttribute="trailing" id="ZLZ-Tf-Sjn"/>
+                            <constraint firstItem="FOa-2C-Uzo" firstAttribute="trailing" secondItem="dBL-fd-hIf" secondAttribute="trailing" constant="15" id="bgc-9e-WId"/>
+                            <constraint firstItem="eg0-fs-3qd" firstAttribute="top" secondItem="Dhn-mi-ndQ" secondAttribute="top" id="mRt-dA-HOb"/>
+                            <constraint firstAttribute="bottom" secondItem="eg0-fs-3qd" secondAttribute="bottom" id="sgV-BL-Vor"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="FOa-2C-Uzo"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="z9m-0E-RCB"/>
+                    <connections>
+                        <outlet property="sceneView" destination="eg0-fs-3qd" id="Rlt-UZ-lyZ"/>
+                    </connections>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="Xao-qS-cu4" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="331.8840579710145" y="-60.267857142857139"/>
+        </scene>
+    </scenes>
+</document>
